[refactor](6차 세미나): viewmodel 사용하여 로그인 통신 구현
